Add Neon implementation of Convert8To16Row
Add a Neon implementation of the Convert8To16Row kernel.

Performance is mostly identical to the existing C code when
auto-vectorized with a recent LLVM, however the assembly is retained as
this still provides consistent performance across a range of compiler
versions and gives us a good basis for building an improved SVE
implementation in a later commit.

Change-Id: I34b858605ece47e46588c0680a1d2afa7a90d7a0