Add Neon implementation of Convert8To16Row
Add a Neon implementation of the Convert8To16Row kernel. Compared to the
C implementation we can take advantage of knowing that the "scale"
parameter is always an unsigned power of two and fits in 16-bits,
allowing us to use widening multiplies rather than needing to separately
widen the input data.

Reduction in run times observed compared to the existing C
implementation:

 Cortex-A55: -12.5%
Cortex-A510: -1.3%
Cortex-A520: -9.6%
 Cortex-A76: -1.9%
Cortex-A710: -15.6%
  Cortex-X1: -15.5%
  Cortex-X2: -19.0%
  Cortex-X3: -20.6%
  Cortex-X4: -10.6%
Cortex-X925: -18.6%

Change-Id: I34b858605ece47e46588c0680a1d2afa7a90d7a0