Revert "Enable explicit control over LoongArch LSX & LASX for GYP builds"
This reverts commit 6ecfe106c304d1eeaa31d1b29a42c58fb70d35ca.

This "explicit control" with -mlasx as the default is really surprising
and unfriendly to distro maintainers and/or downstream projects.  They
already know using -m{no-,}l{a,}sx in CXXFLAGS to enable/disable the
LoongArch vector extensions, but this control does not respect those
flags already in the environment, forcing every distro and every
downstream project to wire an extra logic into their build system if
they need to change the default.

To make things worse, LSX is guaranteed to be available for LoongArch
desktop processors but LASX is not.  Thus most downstream works will
indeed need to change the default for hardware compatibility.

"LASX is widely supported among LoongArch desktops and servers, and
performance is better than with LSX alone" is true, but in the future we
can make the use of LASX correctly guarded by runtime CPU feature check
to take the advantange without breaking the hardware compatibility.

Change-Id: I43b31a139b106b7f6f8b6ca5dc04045633ea1989