Fix compilation issue for 32bit PIC build
Currently, ARGBToUVMatrixRow_AVX2 and ARGBToUVMatrixRow_SSSE3
fail to compile with clang on 32bit PIC build with the error message:
inline assembly requires more registers than available

This is because in PIC code EBX is reserved for the GOT
and with a frame pointer EBP is also unavailable.

Fix this by copying the RGB-to-UV constants to stack locals first and let the
asm use simple stack-relative addressing.

Bug: None
Change-Id: Ica90f0c35039303ecaa145534683f59659fb5d7f